Handle missing docker hub endpoint during maintenance.

+++ scripts/maintenance.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
+
+dockerHubIsUp() {
+  docker pull alpine:3.18 > /dev/null
+}
+
 echo "Updates the host packages ..."
 sudo apt-get clean -y
 sudo apt-get update
@@ -7,25 +12,30 @@ sudo apt-get upgrade -y
 sudo apt-get autoremove -y
 echo "Host packages are updated!"
 
-echo "Stops the containers ..."
-cd /home/ubuntu/home-media-server-pi
-docker-compose --env-file .env.pi down --remove-orphans
-echo "Containers are stopped!"
+echo "Check if DockerHub can be reached ..."
+if ! dockerHubIsUp ; then
+	echo "DockerHub canot be reached, will not update images."
+else
+	echo "Stops the containers ..."
+	cd /home/ubuntu/home-media-server-pi
+	docker-compose --env-file .env.pi down --remove-orphans
+	echo "Containers are stopped!"
 
-echo "Gets update from GitHub ..."
-cd /home/ubuntu/home-media-server-pi
-git pull
-echo "Updates arefetched from GitHub!"
+	echo "Gets update from GitHub ..."
+	cd /home/ubuntu/home-media-server-pi
+	git pull
+	echo "Updates arefetched from GitHub!"
 
-echo "Starts the containers ..."
-cd /home/ubuntu/home-media-server-pi
-docker-compose --env-file .env.pi up -d
-echo "Containers are started!"
+	echo "Starts the containers ..."
+	cd /home/ubuntu/home-media-server-pi
+	docker-compose --env-file .env.pi up -d
+	echo "Containers are started!"
 
-echo "Cleanup images ..."
-cd /home/ubuntu/home-media-server-pi
-docker system prune -af
-echo "Images are cleared!"
+	echo "Cleanup images ..."
+	cd /home/ubuntu/home-media-server-pi
+	docker system prune -af
+	echo "Images are cleared!"
+fi
 
 echo "Checking if reboot is required ..."
 if [ -f /var/run/reboot-required ]; then
